+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Version 0.4.0 (10/27/2016)
+* added `--preview` option to `export disk` command, which opens a simple
+preview map of the exported tiles.
+
+
 ## Version 0.3.0 (10/24/2016)
 * added `path_format` to `TPK::to_disk` and `--path-format` option to 
 CLI `export disk` command to allow control over output filename.